rework config management

+++ src/fuse_ioctl.hpp
@@ -22,7 +22,7 @@ namespace FUSE
 {
   int
   ioctl(const char     *fusepath_,
-        int             cmd_,
+        unsigned long   cmd_,
         void           *arg_,
         fuse_file_info *ffi_,
         unsigned int    flags_,
